remove _common module, put shared methods in ufoLib.__init__
and moved the import ufoLib.glifLib outside from global module scope to
avoid circular import issues
